result float

+++ management/models.py
@@ -144,7 +144,7 @@ class SampleFormParameterFormulaCalculate(models.Model):
     sample_form = models.ForeignKey(SampleForm,related_name="result",on_delete=models.CASCADE,null=True)
     commodity = models.ForeignKey(Commodity,on_delete=models.CASCADE,null=True)
     parameter = models.ForeignKey(TestResult, on_delete=models.CASCADE,null=True)
-    result =  models.IntegerField(null=True)
+    result =  models.FloatField(null=True)
     is_verified = models.BooleanField(default=False)
     input_fields_value = models.CharField(max_length=2000,null=True)
     auto_calculate_result = models.CharField(max_length=200,null=True)
